requirements/common.txt
Flask-Cors>=3.0.10
Flask-Migrate>=3.1.0
Flask-Script>=2.0.6
Flask-SQLAlchemy>=2.5.1
Flask>=2.0.2
PyJWT>=2.3.0
SQLAlchemy>=1.4.26
alembic>=1.7.4
marshmallow>=3.14.0
autopep8>=1.7.0
wheel>=0.37.1
PyMySQL>=1.0.2
psycopg2>=2.9.3
Gunicorn
Flask-OAuthlib>=0.9.6
redis>=4.3.4
python-dotenv>=0.21.0
pre-commit>=2.20.0
blinker>=1.5
flask-dance>=6.2.0